+++ hash/src/ascon_avx512.c
@@ -100,8 +100,8 @@ static void ascon_squeeze(void *state, uint8_t *digest)
 			     ascon_avx512_permutation);
 }
 
-static const struct lc_hash _ascon_128_avx512 = {
-	.init = ascon_128_init,
+static const struct lc_hash _ascon_256_avx512 = {
+	.init = ascon_256_init,
 	.update = ascon_absorb,
 	.final = ascon_squeeze,
 	.set_digestsize = NULL,
@@ -114,7 +114,7 @@ static const struct lc_hash _ascon_128_avx512 = {
 	.statesize = sizeof(struct lc_ascon_hash),
 };
 LC_INTERFACE_SYMBOL(const struct lc_hash *,
-		    lc_ascon_128_avx512) = &_ascon_128_avx512;
+		    lc_ascon_256_avx512) = &_ascon_256_avx512;
 
 static const struct lc_hash _ascon_128a_avx512 = {
 	.init = ascon_128a_init,
@@ -147,19 +147,3 @@ static const struct lc_hash _ascon_xof_avx512 = {
 };
 LC_INTERFACE_SYMBOL(const struct lc_hash *,
 		    lc_ascon_xof_avx512) = &_ascon_xof_avx512;
-
-static const struct lc_hash _ascon_xofa_avx512 = {
-	.init = ascon_xofa_init,
-	.update = ascon_absorb,
-	.final = ascon_squeeze,
-	.set_digestsize = ascon_xof_set_digestsize,
-	.get_digestsize = ascon_xof_get_digestsize,
-	.sponge_permutation = ascon_avx512_permutation,
-	.sponge_add_bytes = ascon_c_add_bytes,
-	.sponge_extract_bytes = ascon_c_extract_bytes,
-	.sponge_newstate = ascon_c_newstate,
-	.sponge_rate = 64 / 8,
-	.statesize = sizeof(struct lc_ascon_hash),
-};
-LC_INTERFACE_SYMBOL(const struct lc_hash *,
-		    lc_ascon_xofa_avx512) = &_ascon_xofa_avx512;