+++ hash/src/ascon_hash_common.h
@@ -26,10 +26,10 @@
 extern "C" {
 #endif
 
-void ascon_128_init(void *_state);
+void ascon_256_init(void *_state);
 void ascon_128a_init(void *_state);
 void ascon_xof_init(void *_state);
-void ascon_xofa_init(void *_state);
+void ascon_cxof_init(void *_state);
 size_t ascon_digestsize(void *_state);
 void ascon_xof_set_digestsize(void *_state, size_t digestsize);
 size_t ascon_xof_get_digestsize(void *_state);